requirements.txt
Django
django-debug-toolbar
djangorestframework
drf-nested-routers
django-filter